+++ components/story-page/back-to-archive-banner/BackToArchiveBanner.styles.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 import { colors } from '../../../style/colors';
 import { P } from '../../../style/typography';
+import { motion } from 'framer-motion';
 
 export const BackToArchiveBannerWrapper = styled.div`
   display: flex;
@@ -22,7 +23,7 @@ export const BackToArchiveText = styled(P)`
   letter-spacing: 0.13em;
 `;
 
-export const BackToArchiveImage = styled.img`
+export const BackToArchiveImage = styled(motion.img)`
   height: 3em;
   width: auto;
   padding-right: 1em;
